Perform checkout and use private route
