+++ client/src/pages/auth/Home.tsx
@@ -1,18 +1,8 @@
 import Layout from '../../components/Layout';
-//import 'bootstrap/dist/css/bootstrap.min.css';
-import NavbarComponent from '../eCom/EComNav';
-import { Container } from 'react-bootstrap';
-import Shop from '../eCom/Shop';
-import CartProvider from '../eCom/CartContext';
 function Home() {
   return (
     <Layout>
-      <CartProvider>
-        <Container>
-          <NavbarComponent></NavbarComponent>
-          < Shop />
-        </Container>
-      </CartProvider>
+      <h1>Home page</h1>
     </Layout>
   );
 }
